server/requirements.txt
asgiref==3.11.1
Django==6.0.3
djangorestframework==3.17.0
gunicorn==25.1.0
packaging==26.0
psycopg==3.3.3
psycopg2-binary==2.9.11
psycopgbinary==0.0.1
sqlparse==0.5.5
typing_extensions==4.15.0
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.6
distro==1.9.0
Django==6.0.3
djangorestframework==3.17.0
google-auth==2.49.1
google-genai==1.70.0
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
numpy==2.4.3
packaging==26.0
pandas==3.0.1
psycopg==3.3.3
psycopg2-binary==2.9.11
psycopgbinary==0.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
